Final cleanup of Shell Navigation tracking for both forward and back navigation
Regression with main page losing itself was happening with back stack and the removed NavigateToPageMessage we had before which was refired on back navigation from the loaded page.
Since we removed that, we now track the frame's navigation in the ShellPage itself as well and use that to set our CurrentPage VM in all cases.
Added some future notes in case we want to remove Frame and manage our own back stack in the future too.