Spec `GoToPageArgs.NavigationMode`
As discussed in #273.

This will allow extensions that want to navigate around the page stack
more control.

For example, the Obsidian Notes extension will want to start with a form
for "Set up vault path". When they submit, instead of just going
home, they'll want to GoToPage('com.obsidian.notes', mode=GoHome), so
that submitting the form takes them to the notes list, rather than just
_home_.

Doesn't actually implement this, of course. Just gets the API ready for
it.