Fix the icons on Window Walker
The ListItems were having their icons manually initialized to `icon`, which was always passed as `string.Empty`.

I suspect this is vestigial from PT Run - where the icon would be paseed in based on the theme. This prevented us from being able to fall back to the icon from the command, which is where we stuck the actual icon