Add 3 settings for Clint
Is your name Clint? Do you like settings? Did you hate that backspace on the
main page could dismiss the palette? If you said yes to any of that, you're in
luck! This PR is for you!

This PR adds three new settings:

* `BackspaceGoesBack`: Controls if backspace with an empty search takes you back
  a level
* `SingleClickActivates`: Allows users to toggle between the "single click
  selects, double activates" vs "single click activates" behavior
* `HighlightSearchOnActivate`: When `true` (**default**), this will select the
  text in the search bar when the window is summoned

This drive-by fixes a bug where clicking to select an item stole focus from the
search box.

closes #306